dump_session: add byref
Setting byref=True will preprocess the module by groveling
through its top-level objects for things that are identical to those in
the other available modules. If it finds matches, it records the names
and discards the objects during pickling.
On load, it then recovers the objects by "import x as y" using the
stored x, y.
This fixes issues with nasty unpickleable things in the depths of
libraries, as well as making the sessions smaller.

Fixes #79 and most of #78.