Fixes some bugs when using `dump_session()` with `byref=True` (#463)
* fix `dump_session(byref=True)` bug when no objetcts are imported from modules

When no objects are found to be imported from external modules, `_stash_modules()` should return `main_module` unmodified, and not the pair list of objects created from it (`original`).

* fix `dump_session(byref=True)` bug when the `multiprocessing` module was imported

Just calling `import multiprocessing` creates a `'__mp_main__'` entry in `sys.modules` that is simply a reference to the `__main__` module. In the old version of the test for module membership, objects in the global scope are wrongly attributed to this virtual `__mp_main__` module. And therefore `load_session()` fails. Testing for object identity instead of module name resolves the issue.

* Save objects imported with an alias and top level modules by reference in `dump_session(byref=TRUE)`

Currently, `dump_session(byref=True)` misses some imported objects. For example:

- If the session had a statement `import numpy as np`, it may find a reference to the `numpy` named
as `np` in some internal module listed in `sys.resources`.  But if the module was imported with a
non-canonical name, like `import numpy as nump`, it won't find it at all. Mapping the objects by id
in `modmap` solves the issue.  Note that just types of objects usually imported under an alias must
be looked up by id, otherwise common objects like singletons may be wrongly attributed to a module,
and such reference in the module could change to a different object depending on its initialization
and state.

- If a object in the global scope is a top level module, like `math`, again `save_session` may find
a reference to it in another module and it works. But if this module isn't referenced anywhere else,
it won't be found because the function only looks for objects inside the `sys.resources` modules and
not for the modules themselves.

This commit introduces two new attributes to session modules saved by reference:
- `__dill_imported_as`: a list with (module name, object name, object alias in session)
- `__dill_imported_top_level`: a list with (module name, module alias in session)

I did it this way for forwards (complete) and backwards (partial) compatibility.

Oh, and I got rid of that nasty `exec()` call in `_restore_modules()`! ;)

* Deal with top level functions with `dump_session()`

Fixes RecursionWarning error where a function defined in the top level of the module being saved with
`dump_session(byref=True)`, of which "globals" is a reference to the original module's `__dict__`,
makes dill to try to save this instead of the modified module's `__dict__`, triggering recursion.

* Added tests for load_session() and dump_session()

* fix singleton comparison, must be by identity, not by equality

* split tests to different files to better test session use cases

* Fix error Py2.7 and Py3.7 where there is a tuple in sys.modules for some reason

* dump_session(): extra test for code coverage

* dump_session and load_session: some minor improvements

* dump_session(): more tests

* dump_session(): dump modules other than __main__ by reference

* dump_session(): minor code coverage investigation

* dump_session() tests: adjustments

* dump_session() tests: fix copyright notice

* dump_session() tests: merge test files using subprocess to test loading in a new session

* tests: Revert change. Test files are independent, should run in any order

* dump_sessio() tests: use an unpickleable object available in PyPy