A hook that allows dill to save the versions of Python and dill
This may be useful later in preventing issues like #488 from occuring in 
the future. By saving the version of Python and dill in the pickle, we 
can save some trouble later. I am also OK with removing the dill 
version, since, with careful planning, we can make sure that we never 
need to use it. This will add at least 50 bytes to very pickle, so maybe 
an option to turn it off would be desirable. I don't know. Thoughts?